+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Fix error when downloading a borehole profile.
 - Fix bug where a facies description with a height of 0 could be created.
 - Fix broken layout of _lithological description_ and _facies description_ columns when a gap of more than one layer was created in the _lithology_ column.
+- Fix bug where lithological descriptions where not copied when copying boreholes or stratigraphies.
 
 ## v2.0.186 - 2023-02-13